Elastic Health Monitor
VB.net app that calls URL of elastic DB with health call and parses the
returned JSON to look at the status of the cluster..  if not green then
it alerts you via email.

Quite simple, with not a lot of error reporting in there as wanted
something quick to monitor and alert me of the status of my cluster.